Adiciona middleware de autenticação por Oauth2 com JWT

+++ app/modules/customer/infrastructure/http/controller/customer_controller.py
@@ -1,6 +1,7 @@
 from typing import Annotated, List, Union
 from http import HTTPStatus
-from fastapi import APIRouter, Depends, Query, Path, HTTPException, Response
+from fastapi import APIRouter, Depends, Query, Path, HTTPException, Response, Request, Security
+from fastapi.security import HTTPBearer
 from fastapi.responses import JSONResponse
 from modules.core.helpers.config_logger import get_logger
 from modules.customer.application.service.customer_service import CustomerService
@@ -9,17 +10,21 @@ from modules.customer.domain.entity.customer_filtered import CustomerFiltered
 
 logger = get_logger()
 
+bearer_scheme = HTTPBearer()
+
 router = APIRouter(
     prefix="/api",
     tags=["Customers"]
 )
 
 
-@router.get("/customer",
-            description="Get all customers",
-            summary="Get all customers",
-            response_model=List[CustomerFiltered]
-            )
+@router.get(
+    path="/customer",
+    description="Get all customers",
+    summary="Get all customers",
+    response_model=List[CustomerFiltered],
+    dependencies=[Security(bearer_scheme)]
+)
 async def get_customers(
     customer_service: Annotated[CustomerService, Depends(CustomerService)],
     page: int = Query(1, ge=1, title="Page Number", description="Page number"),
@@ -38,14 +43,16 @@ async def get_customers(
         )
 
 
-@router.get("/customer/{customer_id}",
-            description="Get customer by id or null",
-            summary="Get customer by id",
-            response_model=CustomerFiltered
-            )
+@router.get(
+    path="/customer/{customer_id}",
+    description="Get customer by id or null",
+    summary="Get customer by id",
+    response_model=CustomerFiltered,
+    dependencies=[Security(bearer_scheme)]
+)
 async def get_customer_by_id(
-        customer_service: Annotated[CustomerService, Depends(CustomerService)],
-        customer_id: str = Path(title="Customer Id", description="Customer Id")
+    customer_service: Annotated[CustomerService, Depends(CustomerService)],
+    customer_id: str = Path(title="Customer Id", description="Customer Id")
 ) -> JSONResponse:
     try:
         return await customer_service.get_by_id(customer_id)
@@ -63,43 +70,55 @@ async def get_customer_by_id(
         )
 
 
-@router.post("/customer",
-            status_code=HTTPStatus.CREATED,
-            description="Create a new customer or fail",
-            summary="Create a new customer",
-            response_model=CustomerFiltered
-            )
+@router.post(
+    path="/customer",
+    status_code=HTTPStatus.CREATED,
+    description="Create a new customer or fail",
+    summary="Create a new customer",
+    response_model=CustomerFiltered,
+    dependencies=[Security(bearer_scheme)]
+)
 async def create_customer(
-        customer_service: Annotated[CustomerService, Depends(CustomerService)],
-        customer: Customer
-) -> JSONResponse:
-    # try:
-        return await customer_service.store(customer)
-    # except HTTPException as ehttp:
-    #     logger.error(f"Error creating customer: {ehttp}")
-    #     return JSONResponse(
-    #         status_code=ehttp.status_code,
-    #         content={"error": ehttp.detail}
-    #     )
-    # except Exception as e:
-    #     logger.error(f"Error creating customer: {e}")
-    #     return JSONResponse(
-    #         status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-    #         content={"error": "Internal Server Error"}
-    #     )
-
-
-@router.put("/customer/{customer_id}",
-            description="Update a customer by Id or fail",
-            summary="Update a customer by id",
-            response_model=CustomerFiltered
-            )
-async def update_customer(
-        customer_service: Annotated[CustomerService, Depends(CustomerService)],
-        customer_id: str,
-        customer: CustomerUp
+    request: Request,
+    customer_service: Annotated[CustomerService, Depends(CustomerService)],
+    customer: Customer
 ) -> JSONResponse:
     try:
+        if request.state.user.profile != "admin":
+            raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail="Forbidden")
+
+        return await customer_service.store(customer)
+    except HTTPException as ehttp:
+        logger.error(f"Error creating customer: {ehttp}")
+        return JSONResponse(
+            status_code=ehttp.status_code,
+            content={"error": ehttp.detail}
+        )
+    except Exception as e:
+        logger.error(f"Error creating customer: {e}")
+        return JSONResponse(
+            status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
+            content={"error": "Internal Server Error"}
+        )
+
+
+@router.put(
+    path="/customer/{customer_id}",
+    description="Update a customer by Id or fail",
+    summary="Update a customer by id",
+    response_model=CustomerFiltered,
+    dependencies=[Security(bearer_scheme)]
+)
+async def update_customer(
+    request: Request,
+    customer_service: Annotated[CustomerService, Depends(CustomerService)],
+    customer_id: str,
+    customer: CustomerUp
+) -> JSONResponse:
+    try:
+        if request.state.user.profile != "admin":
+            raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail="Forbidden")
+
         return await customer_service.update(customer_id, customer)
     except HTTPException as ehttp:
         logger.error(f"Error updating customer: {ehttp}")
@@ -115,16 +134,22 @@ async def update_customer(
         )
 
 
-@router.delete("/customer/{customer_id}",
-                description="Delete a customer or fail",
-                summary="Update a customer",
-                response_model=None
-                )
+@router.delete(
+    path="/customer/{customer_id}",
+    description="Delete a customer or fail",
+    summary="Update a customer",
+    response_model=None,
+    dependencies=[Security(bearer_scheme)]
+)
 async def delete_customer(
+    request: Request,
     customer_service: Annotated[CustomerService, Depends(CustomerService)],
     customer_id: str
 ) -> Union[Response, JSONResponse]:
     try:
+        if request.state.user.profile != "admin":
+            raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail="Forbidden")
+
         await customer_service.delete(customer_id)
         return Response(status_code=HTTPStatus.NO_CONTENT)
     except HTTPException as ehttp:
